requirements.txt
unidecode
django-cms